newest reliquary code

+++ contracts/interfaces/INFTDescriptor.sol
@@ -2,8 +2,5 @@
 pragma solidity ^0.8.15;
 
 interface INFTDescriptor {
-    function constructTokenURI(uint256 relicId)
-        external
-        view
-        returns (string memory);
+    function constructTokenURI(uint relicId) external view returns (string memory);
 }
